\name{ALAE600}
\alias{alae600}
\title{ALAE-loss subset of size 600}
\usage{
data(alae600)
}
\description{
Subset of data set used in "Frees EW and Valdez EA (1998). 
Understanding relationships using copulas.
North American Actuarial Journal, 2, 1-25." 
The subset has 600 uncensored observations.
}
\keyword{datasets}
