\name{asianwklgret}
\alias{asianwklgret}
\alias{hksikotw}
\title{Weekly returns from 4 Asian markets}
\usage{
data(asianwklgret)  # object hksikotw
}
\description{
Weekly returns from 4 Asian markets, July 1997 to May 2006; n=465 weeks.
The indexes are:
Hong Kong Hang Seng (HSI), Singapore Straights (STI), Seoul Kospi (KS11), and 
Taiwan Taiex (TWII). 
Original source is http://quote.yahoo.com
}
\format{
A data matrix with 465 observations on the following 4 variables.
\describe{
\item{\code{hksikotw}}{columns are hk, si, ko, tw}
}
}
\keyword{datasets}
