\name{kzrepmeas}
\alias{kzrepmeas}
\title{Karim-Zeger data set}
\usage{
data(kzrepmeas)
}
\description{
73 subjects, with 4 repeated measures in quarters 1 to 4;
variables are id, time, age, sex, msmok, y. For the variables,
y is the count response variable for the number of visits to the hospital,
quarterly over a one-year period;
id=identification code, time=quarter, 
age=age in months at the beginning of the study,
sex=0 for male and 1 for female,
msmok=maternal smoking status (1 for yes, 0 for no).

This data set is on 
page 342, Table 9.22, Problem 9.12 of
"Statistical methods for the analysis of repeated measurements",
by Charles S. Davis.

The original source is:
"Karim MR and Zeger SL (1988). GEE: A SAS macro for longitudinal
data analysis. Technical Report 674, Department of Biostatistics, The
Johns Hopkins University, Baltimore."
}
\keyword{datasets}
