\name{ltmconv}
\alias{ltmconv}
\alias{sci}
\alias{env}
\title{Item response data sets from ltm R package}
\usage{
data(ltmconv) # objects are sci and env
}
\description{
item response data sets converted from ltm R package:
(a) science, (b) environment.

(a) sci: Science data set, size 392x7 (7 items, categories 0,1,2,3);
"0=strongly disagree", "1=disagree to some extent", "2=agree to some
extent" and "3=strongly agree".
Sample size n=392.
The items are:

Y1: Science and technology are making our lives healthier, easier and
more comfortable;

Y2: Scientific and technological research cannot play an important role
in protecting the environment and repairing it;

Y3: The application of science and new technology will make work more
interesting;

Y4: Thanks to science and technology, there will be more opportunities
for the future generations;

Y5: New technology does not depend on basic scientific research;

Y6: Scientific and technological research do not play an important role
in industrial development;

Y7: The benefits of science are greater than any harmful effect it may have.

(b) env: Environment data set, size 291x6 (6 items, categories 0,1,2).
"0=very concerned", "1=slightly concerned" and "2=not very concerned".
Sample size n=291.
The items are:

Y1: Lead from petrol;

Y2: River and sea pollution;

Y3: Transport and storage of radioactive waste;

Y4:  Air pollution;

Y5: Transport and disposal of poisonous chemicals;

Y6: Nuclear Risks from nuclear power station.

}
\keyword{datasets}
