\name{rwmsubset}
\alias{rwmsubset}
\title{Riphahn-Wambach-Million subset}
\usage{
data(rwmsubset)
}
\description{
50 subjects, with 5 repeated measures in years 1984, 1985, 1986, 1987,
1988 (5 lines per subject); variables are
id, docvis, sex, year, age, hsat, handper, univ, public,
addon, married, self.

This data set is a subset (of cases and variables) from
"Riphahn RT, Wambach A, and Million A (2003). Incentive Effects in
the Demand for Health Care: A Bivariate Panel Count Data Estimation.
Journal of Applied Econometrics, 18 (4), 387-405."
}
\keyword{datasets}
