\name{wcb8594}
\alias{wcb}
\alias{wcb8594}
\title{Workmans' Compensation Board monthly claims 1985-1994}
\usage{
data(wcb8594)
}
\description{
Data set from "Freeland RK (1998), Statistical Analysis of Discrete Time Series
with Application to the Analysis of Workers' Compensation Claims Data,
PhD thesis, University of British Columbia."

variable 1: nclaims is the
monthly claims of short-term disability benefits made by injured workers
from 1985 to 1994; category is cut injury in the logging industry.

variable 2: x1 has sin(2*pi*t/12), t=1,...,120.

variable 3: x2 has cos(2*pi*t/12), t=1,...,120.

x1 and x2 are considered as vectors of seasonal covariates.
}
\keyword{datasets}
